refactor(calculation): 패키지명을 미션 주제에 맞는 이름으로 변경
