+++ src/main/kotlin/study/calculation/Calculation.kt
@@ -1,4 +1,4 @@
-package study.step2
+package study.calculation
 
 interface Calculation {
     fun calculate(a: Int, b: Int): Int
