+++ src/main/kotlin/study/calculation/Multiplication.kt
@@ -1,4 +1,4 @@
-package study.step2
+package study.calculation
 
 class Multiplication : Calculation {
     override fun calculate(a: Int, b: Int) = a * b
