+++ src/main/kotlin/study/calculation/Subtraction.kt
@@ -1,4 +1,4 @@
-package study.step2
+package study.calculation
 
 class Subtraction : Calculation {
     override fun calculate(a: Int, b: Int) = a - b
